Dynamically export bcftools symbols to plugins
...instead of linking the plugins against libhts.so.  This simplifies
building and installation, as plugins no longer need to be able to find
libhts.so/.dylib at runtime -- which to date has unfortunately often
been done via $LD_LIBRARY_PATH.

On OS X, build plugins as Mach-O bundles rather than shared libraries
(and ensure that bcftools is built first, so is available for use with
-bundle_loader).  Fixes #225.

(TODO) The advice in print_plugin_usage_hint() should be revisited.

When bcftools is linked to libhts.a statically, there is a small chance
that a plugin will need HTSlib functions that have not been pulled in by
the bcftools executable.  This is surely unlikely, and can be prevented
by linking bcftools or your plugin against a shared libhts.so.

Fixes #249; fixes #224; supersedes and closes #254.