Improvements in "bcftools call --ploidy"
Previously samples of unknown sex were assigned the
default ploidy of 2 and the only way to override the
default was to create a sample file. This is not required
any more, one can assign the default ploidy for all
sexes using the "* * * * <int>" rule.

If all samples are haploid, run with "--ploidy 1".

Resolves #366.